LUXEN-22 Read MySQL configuration from file
Until now MySQL config was hardcoded. This commit reads the
configuration from the server.conf file (which we already use
for logging levels), which also means we need to use a ring
middleware to add the DB connection to the request (as app
state) so that we can pass it to the handler.